refactor: create custom hooks for cards and detail, refactor pagination function for prev and next
